Create make_scrobble function
Generalise the creation of parameter lists for last.fm requests or for
storing failed scrobbles. Takes the value from the player status and
song, as well as optional values to place in the returned dict too. If
an api secret is given, sign the scrobble too.

This can reduce the amount of repetitive code in other functions
(now_playing, send_scrobble) and make it easier to add additional
properties as needed in future.