Use make_scrobble function in scrobble functions
Use the new make_scrobble function to replace the parameter creation
functionality in the now_playing and scrobble_track. As very little code
remained in the record_failed_scrobble function after this change, I've
deleted it and simply placed the check for whether the failed scrobble
already exists in the two places it's needed.