Use arbitrary kwargs in make_scrobble()
Rather than specifying all the possible parameters that can be added to
a scrobble, instead take any provided key-value pairs and add it to the
scrobble before signing.

This means that argument names have to match the parameter name in the
last.fm api, which is why the change from session_key to sk is needed.