Refactor file handling to use platformdirs for config, state files etc.
Instead of hard coding the config directory use platformdirs to find
the configuration directory. The benefit is that if the user has
changed XDG_CONFIG_HOME to a different directory the program still
works.
The side effect is that the program is more portable now.

Apply this also to the session file, the log file and the pid file to
move them to their appropriate directories. Fixes #17.

Requires regeneration of config file as these are stored in the
configuration file.

Includes #21.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>